message-admin-rest/requirements.txt
flask
flask-restful
flask-sqlalchemy
sqlalchemy-utils
marshmallow-sqlalchemy
flask_marshmallow
pytest
pytest_mockito
pytest-flask